Move theme check
