+++ frontend/eslint.config.js
@@ -6,7 +6,7 @@ import globals from 'globals';
 import { fileURLToPath } from 'node:url';
 import ts from 'typescript-eslint';
 import svelteConfig from './svelte.config.js';
-import themeCheck from './eslint-theme-check.js';
+import themeCheck from './eslint-rules/eslint-theme-check.js';
 import noDirectDateConversion from './eslint-rules/no-direct-date-conversion.js';
 
 const gitignorePath = fileURLToPath(new URL('./.gitignore', import.meta.url));
@@ -35,7 +35,7 @@ export default ts.config(
       // see: https://typescript-eslint.io/troubleshooting/faqs/eslint/#i-get-errors-from-the-no-undef-rule-about-global-variables-not-being-defined-even-though-there-are-no-typescript-errors
       'no-undef': 'off',
       // Enable our custom theme checking rule
-      'theme-check/no-hardcoded-colors': 'warn',
+      'theme-check/no-hardcoded-colors': 'error',
       // Enable our custom date handling rule
       'custom/no-direct-date-conversion': 'error',
     },
